# services/booking-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para reservas, asientos, checkout y tickets.",
        "title": "SeatGuard Booking Service API",
        "contact": {
            "name": "Lucas Cabral",
            "email": "lucassimple@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/seats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener todos los asientos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Obtener asientos",
                "responses": {
                    "200": {
                        "description": "Asientos obtenidos satisfactoriamente",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Seat"
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los asientos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crear un nuevo asiento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Crear asiento",
                "parameters": [
                    {
                        "description": "Datos del asiento",
                        "name": "seat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Asiento creado satisfactoriamente",
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    },
                    "400": {
                        "description": "Formato JSON inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el asiento",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seats/event/{eventId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener todos los asientos por ID de evento",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Obtener asientos por ID de evento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del evento",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asientos obtenidos satisfactoriamente",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Seat"
                            }
                        }
                    },
                    "400": {
                        "description": "Formato UUID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Asientos no encontrados",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los asientos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seats/lock/{id}/uid/{uid}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bloquear asiento por 15 minutos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Bloquear asiento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del asiento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UID del usuario",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asiento bloqueado satisfactoriamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Formato UUID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Asiento no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error al bloquear el asiento",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seats/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener un asiento por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Obtener asiento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del asiento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asiento obtenido satisfactoriamente",
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    },
                    "400": {
                        "description": "Formato UUID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Sin resultados",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el asiento",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualizar el estado de un asiento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seats"
                ],
                "summary": "Actualizar estado de asiento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del asiento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado del asiento",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asiento actualizado satisfactoriamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Formato UUID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Asiento no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el asiento",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Seat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventId": {
                    "description": "Relaciones",
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lockedAt": {
                    "description": "Cuando se desbloquea",
                    "type": "string"
                },
                "lockedBy": {
                    "description": "Control de Bloqueo Temporal",
                    "type": "string"
                },
                "number": {
                    "description": "\"10\", \"A1\", etc.",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "section": {
                    "description": "Identificación del asiento (ej: A-10, B-5)",
                    "type": "string"
                },
                "status": {
                    "description": "Estado actual del asiento\nenums: AVAILABLE, RESERVED, SOLD, BLOCKED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SeatStatus"
                        }
                    ]
                },
                "ticketId": {
                    "description": "ID del ticket final si se vende",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SeatStatus": {
            "description": "Estado del asiento",
            "type": "string",
            "enum": [
                "AVAILABLE",
                "LOCKED",
                "SOLD"
            ],
            "x-enum-varnames": [
                "StatusAvailable",
                "StatusLocked",
                "StatusSold"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Formato: Bearer \u003ctoken\u003e\nOperaciones para gestionar eventos\nOperaciones para gestionar asientos\nOperaciones para gestionar órdenes de reserva\nOperaciones para gestionar el proceso de checkout",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}